Fixed some of the include errors.
Fixed up the use of sessions - seemed to be a misunderstanding of how sessions are created differently to cookies. It is general practice to just have one session to track a browser which uses one session id cookie in the client.  The session is created as soon as the browser connects, before we login, not at the time of authentication.  So we can't just check for the existence of the session - we need to check it's content to make sure it has what we want.