Allowed the triggering of workflows from the 'My Fez' page.  These workflows are special because they require the user to select the PIDs for the records to be associated with the workflow after the workflow has been started.  (We start them initially with PID = -2)
Changes to workflow so that each running workflow has a unique ID, not based on PID as sometimes the pid is not known at the start of the workflow.

The form for selecting the records at the start of the workflow uses NAJAX (a PHP AJAX protocol lib) to dynamically load the select boxes as the user selects the community and collection.