Speed up auth index #229485 Implemented rule groups and rule group users.  Groups of rules are common to many pids through inheritance.
Also, when a user logs in, they are matched against rule groups and the results kept in a table.  When doing queries we can join on the pids that have rule groups that the user matched.

When saving the security, we try to cache as much as possible while recursing and also don't recurse if we can determine that nothing has changed.