Implemented a way of setting messages for the user from deep within code without necessarily logging an error.  This is used for when we want to tell the user that their file upload didn't work.
Also changed the file upload thing to discard the mimetype sent by the browser as it is unreliable.  instead we try to detect the mimetype using the 'file' unix command on the fez server (or the built in php mime_content_type if it is available).

Added ordering back into the controlled vocab selector

change fulltext index so that it deletes entries when purging pids or datastreams