Support for indexing files from fedora where the datastream doesn't have a file extension on the image.  It all works off the mimetype and had to add checks everywhere for assumed extension on these things!.
NOTE / TODO: I hope i haven't broken batchimport and eprints import doing this.  The presmd thing used to just reject all files with no extension - i assume so that the xml datastreams wouldn't be presmd'd.  I'm not sure if batchimporting is able to check before it calls the presmd stuff.